+++ src/main/java/com/example/gatherplan/appointment/mapper/AppointmentMapper.java
@@ -1,14 +1,8 @@
 package com.example.gatherplan.appointment.mapper;
 
-import com.example.gatherplan.api.vo.DailyWeatherClientResp;
-import com.example.gatherplan.api.vo.KeywordPlaceClientResp;
 import com.example.gatherplan.appointment.dto.CreateAppointmentReqDto;
-import com.example.gatherplan.appointment.dto.DailyWeatherRespDto;
-import com.example.gatherplan.appointment.dto.KeywordPlaceRespDto;
-import com.example.gatherplan.appointment.dto.RegionDto;
 import com.example.gatherplan.appointment.enums.AppointmentState;
 import com.example.gatherplan.appointment.repository.entity.Appointment;
-import com.example.gatherplan.appointment.repository.entity.Region;
 import org.mapstruct.*;
 
 @Mapper(
@@ -23,10 +17,4 @@ public interface AppointmentMapper {
     @Mapping(target = "id", ignore = true)
     @Mapping(target = "confirmedDateTime", ignore = true)
     Appointment to(CreateAppointmentReqDto req, AppointmentState appointmentState, String appointmentCode);
-
-    RegionDto to(Region entity);
-
-    KeywordPlaceRespDto to(KeywordPlaceClientResp.KeywordPlaceInfo clientResp);
-
-    DailyWeatherRespDto to(DailyWeatherClientResp.DailyWeatherInfo clientResp);
 }
