+++ src/main/java/com/example/gatherplan/appointment/service/AppointmentService.java
@@ -1,16 +1,8 @@
 package com.example.gatherplan.appointment.service;
 
-import com.example.gatherplan.appointment.dto.*;
-
-import java.util.List;
+import com.example.gatherplan.appointment.dto.CreateAppointmentReqDto;
 
 public interface AppointmentService {
-
-    List<RegionDto> searchRegion(RegionReqDto reqDto);
-
-    List<KeywordPlaceRespDto> searchKeywordPlace(KeywordPlaceReqDto reqDto);
-
-    List<DailyWeatherRespDto> searchDailyWeather(DailyWeatherReqDto reqDto);
 
     String registerAppointment(CreateAppointmentReqDto reqDto, String email);
 }
